Game Template Running (Read Description)
This version of the project is up and running with Area inclusion functionality. It does not have tiles, players, enemies, weapons, or items, but it DOES have backgrounds and area loading working. 

Listed below is the folder hierarchies for the classes.

Standard files that came with cocos2d are kept in their respective locations.
- Imports of our OOP Labs are just kept in the Classes folder, with no sub folders used. They are all part of the OOP namespace.
- Entity and entity children are kept in the folder 'entities'.
- Area classes are kept in the 'areas' subfolder. Areas that fall under a specific location are kept in a folder named after said place (e.g. "AIN_X00.cpp" is stored in folder 'AIN_X'.
-All user made scenes are kept in the 'scenes' subfolder, and should be given the 'MSQ_' prefix.

Listed below are the Global Z Orders for current sprites.
BG1 -> 0.0F
BG2 -> 0.1F
BG3 -> 0.2F
FG -> 9.0F
Mouse Label -> 10.0F
Grid -> 10.1F

Any and all questions should be directed at the committer of this message regarding this setup.