Main Template Near Completion
The main template for the program is near completion. All that's left is to add collision, scene switching, camera scrolling, and types for the base game. You can download this for now, but there are some image files that need to be changed, which will be dealt with at a later time. If you need them immediately, just ask me (Roderick).