Project Build Version 1.9
This is a working build of the project. Various elements of the template are in place, but not fully complete. You all can at least start working with it and getting a feel for it if you want. 

Also, important note. Even though the PlatformManager is included in this build, it crashes the build. DO NOT ADD THE PLATFORM_MANAGER TO YOUR SOLUTION EXPLORER. IT WILL CRASH. I don't know why this happens, but I'm going to see if I can fix it later, even if we don't end up using it in the final game.