Audio stuff
Still not 100% sure if it works, and a few things will need changing, but I thought that the sooner this was up, the better. Before it's implemented I'll make sure all the audio files are pushed for everyone and that the right audio files are actually available. This being added won't affect anything, it does nothing as of right now. It's 5:33am and I haven't slept so I might be repeating myself, but I just want to make sure my message gets across that this won't do anything, I just think the files being pushed will get the file structure solid asf :ok_hand: