Saving Current Progress ~ Please Read
I was trying to get scene switching working, but I've caught a snag where it crashes once the transition is finished. However, I've decided to push what I currently have.

I have included the MyArray lab from last week, though I don't think we'll use it. I also created an AudioLibrary to house all of our sound effects, and put them in an audio subfolder. I've also added a transition scene in an attempt to get things working, but whether or not we need it remains to be seen. Oh, and I've also removed the physics world stuff since we aren't using it.