Scene Transitions!... with some Caveats
Scene transitions now work. However, if this happens, we can't use transitions because then the player can't move until the transition finishes, which could cause unfair consequences. As such, I've set it not to use any transitions.

Oh, I've also added a bool called ENABLE_CAMERA to toggle on and off the camera so that you guys can use that for testing.