Weapon Collisions ~ Auto stash before merge of "master" and "origin/master"
Weapon collisions have been added! It's just one hit box unfortunately, but I think that's for the best given the time constraints. The attack has also been changed from happening when the key is pressed versus when the key is up. Finally, I have made modifications to the physics that I intend to add a toggle tomorrow or so.

It's kind of ugly, but I'll clean it up tomorrow.