Minecraft 1.69: Main Menu Update (pls read description for changes)
- Added enemy placements in F01 and fixed some bits of F00. 
- Fixed the spawn location in X00 so you don't spawn mid-air and just fall
- Added the Main Menu and support for switching scenes to anything really
- Changed the attack sound effect
- Removed the W and UpArrow to move up and S and DownArrow to move down thing
- Changes to AppDelegate.cpp that allow the MainMenu to be "played" (literally a screen with text and two button prompts, one of which closes the game and the other of which crashes the game)
- Potential FPS counter in GameplayScene

So yeah, I've been busy. Little side note, the Main Menu needs to be manually switched back to in the AppDelegate.cpp, just cause I didn't want to screw around with anyone working on stuff in the GameplayScene. On release we'll change that, obviously, but yeah. 
Enjoy this BS and I'll do more fixes to this stuff in the morning, I'm just hella tired. Goodnight.